requirement.txt
Django==1.5.4
Markdown==2.3.1
MySQL-python
Pillow==2.2.1
South==0.8.2
Unidecode==0.04.14
XlsxWriter==0.5.2
argparse==1.2.1
django-autoslug==1.7.1
django-extensions==1.2.5
django-filer==0.9.5
django-filter==0.7
django-guardian==1.1.1
django-js-reverse==0.2.1
django-mptt==0.5.5
django-nvd3==0.6.0
django-object-log==0.7
django-polymorphic==0.5.3
django-selectable==0.8.0
django-taggit==0.10
djangorestframework==2.3.8
easy-thumbnails==1.4
ipdb==0.8
ipython==1.1.0
progressbar
python-dateutil==2.2
python-nvd3==0.11.0
pytz==2013.8
six==1.4.1
sqlparse==0.1.10
wsgiref==0.1.2
